backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
supabase>=2.5.0
razorpay>=1.4.1
google-genai>=0.1.2
python-dotenv>=1.0.1
pydantic>=2.7.4
pydantic-settings>=2.0.0
python-multipart>=0.0.9
